src/backend/ml_api/requirements.txt
transformers==4.34.0
torch==2.5.1
requests==2.31.0
sentencepiece==0.1.97
gunicorn==21.2.0
protobuf==4.24.4
pytest==8.3.3
openai==1.55.1
scikit-learn==1.3.2
numpy==1.24.4
pyjwt==2.10.1
fastapi[standard]==0.115.6